Refactor analysis and scheduling utilities
