+++ python/tvm/dlight/__init__.py
@@ -17,12 +17,14 @@
 """DLight package provides efficient schedules out-of-box for deep learning workloads."""
 from . import gpu
 from . import cpu
-from .base import (
-    ApplyDefaultSchedule,
+from .analysis import (
     BlockInfo,
     IterInfo,
-    ScheduleRule,
     normalize_prim_func,
+)
+from .base import (
+    ApplyDefaultSchedule,
+    ScheduleRule,
     try_inline,
     try_inline_contiguous_spatial,
 )
