+++ python/tvm/dlight/gpu/reduction.py
@@ -21,13 +21,13 @@ from typing import List, Mapping, Optional, Tuple, Union
 from tvm import arith, ir, tir
 from tvm.target import Target
 
-from ..base import (
+from ..analysis import (
     BlockInfo,
     detect_dominant_read,
     is_broadcast_epilogue,
     normalize_prim_func,
-    try_inline_contiguous_spatial,
 )
+from ..base import try_inline_contiguous_spatial
 from . import utils
 from .base import GPUScheduleRule
